% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{API_BASE_URL}
\alias{API_BASE_URL}
\title{API Base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
API_BASE_URL
}
\description{
Base URL for the Synthesize Bio API
}
\keyword{datasets}
