% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-hosted.R
\name{arrow_schema_metadata}
\alias{arrow_schema_metadata}
\title{Extract Schema-Level Metadata from an Arrow Table (Internal)}
\usage{
arrow_schema_metadata(table)
}
\arguments{
\item{table}{An Arrow `Table`.}
}
\value{
A named list of metadata strings (possibly empty).
}
\description{
Extract Schema-Level Metadata from an Arrow Table (Internal)
}
\keyword{internal}
