% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-hosted.R
\docType{data}
\name{ARROW_STREAM_CONTENT_TYPE}
\alias{ARROW_STREAM_CONTENT_TYPE}
\title{Arrow IPC Stream Content Type}
\format{
An object of class \code{character} of length 1.
}
\usage{
ARROW_STREAM_CONTENT_TYPE
}
\description{
MIME type used by the self-hosted container for synchronous
Apache Arrow IPC stream responses.
}
\keyword{internal}
