% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_POLL_TIMEOUT_SECONDS}
\alias{DEFAULT_POLL_TIMEOUT_SECONDS}
\title{Default Poll Timeout}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_POLL_TIMEOUT_SECONDS
}
\description{
Default maximum timeout (seconds) for async model queries
}
\keyword{datasets}
