% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{DEFAULT_TIMEOUT}
\alias{DEFAULT_TIMEOUT}
\title{Default Timeout}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
DEFAULT_TIMEOUT
}
\description{
Default timeout (seconds) for outbound HTTP requests
}
\keyword{datasets}
