% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_model_api.R
\name{get_json}
\alias{get_json}
\title{Get JSON from URL}
\usage{
get_json(url)
}
\arguments{
\item{url}{The URL to fetch from}
}
\value{
The parsed JSON content
}
\description{
Internal function to fetch JSON from a URL
}
\keyword{internal}
