% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{MODEL_MODES}
\alias{MODEL_MODES}
\title{Model Modes}
\format{
A nested list with structure: version > modality > mode
}
\usage{
MODEL_MODES
}
\description{
A nested list containing supported modes for different
model versions
+ bulk = bulk RNA-seq
+ single-cell = single-cell RNA-seq
}
\keyword{datasets}
