% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-transformers.R
\docType{data}
\name{OUTPUT_TRANSFORMERS}
\alias{OUTPUT_TRANSFORMERS}
\title{Output Transformer Registry}
\format{
An object of class \code{list} of length 4.
}
\usage{
OUTPUT_TRANSFORMERS
}
\description{
A registry mapping model IDs to their corresponding output transformer functions.
Models not in the registry will use the default standard transformer.
}
\keyword{internal}
