% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_model_api.R
\name{start_model_query}
\alias{start_model_query}
\title{Start Model Query}
\usage{
start_model_query(api_base_url, api_slug, query)
}
\arguments{
\item{api_base_url}{The base URL for the API}

\item{api_slug}{The API slug for the specific model}

\item{query}{The query list}
}
\value{
The model query ID
}
\description{
Internal function to start an async model query
}
\keyword{internal}
