% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{synthesize}
\alias{synthesize}
\title{Synthesize Bio API R Client}
\description{
A package for interacting with the Synthesize Bio API to generate
gene expression data based on specified biological conditions.
}
