% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_model_api.R
\name{transform_result_to_frames}
\alias{transform_result_to_frames}
\title{Transform Result to Frames}
\usage{
transform_result_to_frames(content)
}
\arguments{
\item{content}{The parsed JSON content}
}
\value{
A list with expression and metadata data frames
}
\description{
Internal function to transform the final JSON result into data frames
}
\keyword{internal}
