% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-transformers.R
\name{default_output_transformer}
\alias{default_output_transformer}
\title{Transform Metadata Model Output (Internal)}
\usage{
default_output_transformer(final_json)
}
\arguments{
\item{final_json}{The parsed API response list}
}
\value{
the same final_json
}
\description{
Default output transformer, does not modify the response from the server.
}
\keyword{internal}
