% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{MODEL_MODALITIES}
\alias{MODEL_MODALITIES}
\title{Model Modalities}
\format{
A nested list with structure: model type > version > modalities
}
\usage{
MODEL_MODALITIES
}
\description{
A nested list containing supported modalities for different
model versions
+ sra = this is bulk RNA-seq
}
\keyword{datasets}
